streamlining subscription and mutation configuration in the vertx server
