+++ sqrl-server/sqrl-server-core/src/main/java/com/datasqrl/graphql/server/Context.java
@@ -16,6 +16,4 @@ public interface Context {
   DataFetcher<Object> createPropertyFetcher(String name);
 
   DataFetcher<?> createArgumentLookupFetcher(GraphQLEngineBuilder server, Map<Set<Argument>, ResolvedQuery> lookupMap);
-
-  MutationCoordsVisitor createSinkFetcherVisitor();
 }
