+++ sqrl-server/sqrl-server-core/src/main/java/com/datasqrl/graphql/server/MutationConfiguration.java
@@ -1,0 +1,7 @@
+package com.datasqrl.graphql.server;
+
+import com.datasqrl.graphql.server.RootGraphqlModel.MutationCoordsVisitor;
+
+public interface MutationConfiguration<R> {
+  MutationCoordsVisitor<R, Context> createSinkFetcherVisitor();
+}
