+++ sqrl-server/sqrl-server-vertx/src/main/java/com/datasqrl/graphql/config/ServerConfig.java
@@ -39,4 +39,11 @@ public class ServerConfig {
   CorsHandlerOptions corsHandlerOptions;
   @Nullable
   JWTAuthOptions authOptions;
+
+  // I moved it here as I believe it belongs to the server configuration.
+  // The method itself is needed for easier mocking.
+  // We can consider moving it to real server configuration later.
+  public String getEnvironmentVariable(String envVar) {
+    return System.getenv(envVar);
+  }
 }
